cloud_platform/app/requirements.txt
flask
paho-mqtt==1.6.1